% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian_merge.R
\name{hessian_download}
\alias{hessian_download}
\title{Hessian Gather}
\usage{
hessian_download(working.dir, top.dir = "condor_hessian", ...)
}
\arguments{
\item{working.dir}{local directory containing Hessian subdirectories with
MFCL input files, but not Hessian output files.}

\item{top.dir}{top directory on Condor submitter machine containing parallel
Hessian computations.}

\item{\dots}{passed to \code{\link{condor_download}}.}
}
\value{
Remote directory names with the job id as a name attribute.
}
\description{
Gather parallel Hessian computations from Condor and rename \verb{.hes}
files.
}
\examples{
\dontrun{
session <- ssh_connect("servername")
hessian_download("c:/x/yft/hessian")
}

}
