% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian_input.R
\name{hessian_input}
\alias{hessian_input}
\title{Hessian Input}
\usage{
hessian_input(original.dir, working.dir, njobs, force = FALSE)
}
\arguments{
\item{original.dir}{directory containing a converged model run.}

\item{working.dir}{directory where Hessian subdirectories will be created.}

\item{njobs}{number of parallel jobs to create.}

\item{force}{whether to remove existing subdirectories.}
}
\value{
Names of subdirectories created.
}
\description{
Prepare Hessian input subdirectories for parallel computations.
}
\examples{
\dontrun{
hessian_input("z:/yft/2023/model_runs/diagnostic", "c:/x/yft/hessian", 16)
}

}
