% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian_monolithic.R
\name{hessian_monolithic}
\alias{hessian_monolithic}
\title{Hessian Monolithic}
\usage{
hessian_monolithic(original.dir, working.dir = NULL, overwrite = FALSE,
  quiet = FALSE)
}
\arguments{
\item{original.dir}{directory containing input files for a model run.}

\item{working.dir}{directory where Hessian run will be prepared.}

\item{overwrite}{whether to remove existing working directory.}

\item{quiet}{whether to suppress messages.}
}
\value{
Working directory name where files have been prepared.
}
\description{
Prepare a Hessian monolithic model run by adding Hessian flags to an existing
\verb{doitall.sh} script.
}
\details{
The default value of \code{working.dir = NULL} creates a local working
directory with the same name as \code{original.dir} but with a prefix
\verb{Hess_}.
}
\note{
The term \emph{monolithic} refers to a \verb{doitall.sh} script that first
fits a model and then calculates the Hessian.

A monolithic calculation takes a longer time run than a \verb{stand-alone} or
\emph{parallel} Hessian calculation.
}
\examples{
\dontrun{
hessian_monolithic("z:/yft/2023/model_runs/diagnostic")
}

}
