% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian_serial.R
\name{hessian_serial}
\alias{hessian_serial}
\title{Hessian Serial}
\usage{
hessian_serial(original.dir, working.dir = basename(original.dir))
}
\arguments{
\item{original.dir}{directory containing a converged model run.}

\item{working.dir}{directory where Hessian run will be prepared.}
}
\value{
Files in working directory after preparation.
}
\description{
Prepare working directory for serial Hessian calculation.
}
\examples{
\dontrun{
hessian_serial("z:/yft/2023/model_runs/diagnostic")
}

}
