% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian_split.R
\name{hessian_split}
\alias{hessian_split}
\title{Hessian Split}
\usage{
hessian_split(original.dir, working.dir, njobs, force = FALSE)
}
\arguments{
\item{original.dir}{directory containing a converged model run.}

\item{working.dir}{directory where subdirectories will be created.}

\item{njobs}{number of parallel jobs to create.}

\item{force}{whether to remove existing subdirectories.}
}
\value{
Names of subdirectories created.
}
\description{
Split parallel Hessian computations into subdirectories so they are ready to
submit.
}
\examples{
\dontrun{
hessian_split("z:/yft/2023/model_runs/diagnostic", "c:/x/yft/hessian", 16)
}

}
