% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian_stub.R
\name{hessian_stub}
\alias{hessian_stub}
\title{Hessian Stub}
\usage{
hessian_stub(original.dir, working.dir = basename(original.dir),
  overwrite = FALSE)
}
\arguments{
\item{original.dir}{directory containing a converged model run.}

\item{working.dir}{directory where Hessian run will be prepared.}

\item{overwrite}{whether to remove existing working directory.}
}
\value{
Filenames in working directory after preparation.
}
\description{
Prepare working directory for Hessian calculation, starting from a converged
model run.
}
\note{
The term \emph{stub} refers to a stand-alone calculation of the Hessian,
starting from a converged model.

A stub calculation takes a shorter time to run than a \emph{monolithic} run,
where an extended \verb{doitall.sh} script first fits a model and then
calculates the Hessian.

A stub calculation takes a longer time to run than a \emph{parallel} Hessian
calculation, where computations are run simultaneously on many cores.
}
\examples{
\dontrun{
hessian_stub("z:/yft/2023/model_runs/diagnostic")
}

}
