% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.npar.R
\name{read.npar}
\alias{read.npar}
\title{Read Number of Parameters}
\usage{
read.npar(parfile)
}
\arguments{
\item{parfile}{name of \file{.par} file to read from.}
}
\value{
Number of parameters as an integer.
}
\description{
Read the number of estimated parameters from a \file{.par} file.
}
\examples{
\dontrun{
read.npar("10.par.par")
}

}
