% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npar.R
\name{read_npar}
\alias{read_npar}
\title{Read Number of Parameters}
\usage{
read_npar(parfile)
}
\arguments{
\item{parfile}{name of \file{.par} file to read from.}
}
\value{
Number of parameters as an integer.
}
\description{
Read the number of estimated parameters from a \file{.par} file.
}
\examples{
\dontrun{
read_npar("10.par")
}

}
\seealso{
\code{\link[FLR4MFCL]{read.MFCLPar}} reads all the information contained in a
\file{.par} file, which takes substantially longer.
}
