% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian_run.R
\name{run_hessian}
\alias{run_hessian}
\title{Submit Hessian}
\usage{
run_hessian(working.dir, top.dir = "condor_hessian", ...)
}
\arguments{
\item{working.dir}{local directory containing Hessian subdirectories.}

\item{top.dir}{top directory on Condor submitter machine for running parallel
Hessian computations.}

\item{\dots}{passed to \code{\link{condor_submit}}.}
}
\value{
Type of output.
}
\description{
Submit parallel Hessian computations to Condor.
}
\examples{
\dontrun{
run_hessian("c:/x/yft/hessian")
}

}
