% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.hessian.R
\name{split_hessian}
\alias{split_hessian}
\title{Split Hessian}
\usage{
split_hessian(original.dir, working.dir, njobs, force = FALSE)
}
\arguments{
\item{original.dir}{directory containing a converged model run.}

\item{working.dir}{directory where subdirectories will be created.}

\item{njobs}{number of parallel jobs to create.}

\item{force}{whether to remove existing subdirectories.}
}
\value{
Names of subdirectories created.
}
\description{
Prepare subdirectories for parallel Hessian computations.
}
\examples{
\dontrun{
split_hessian("z:/yft/2023/model_runs/diagnostic", "c:/hessian", njobs=16)
}

}
