% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_split.R
\name{submit_split}
\alias{submit_split}
\title{Submit Split}
\usage{
submit_split(working.dir = ".", top.dir = "condor_hessian", ...)
}
\arguments{
\item{working.dir}{local directory containing Hessian subdirectories.}

\item{top.dir}{top directory on Condor submitter machine for running parallel
Hessian computations.}

\item{\dots}{passed to \code{\link{condor_submit}}.}
}
\value{
Data frame showing submitted jobs.
}
\description{
Submit Hessian split subdirectories to Condor for parallel computations.
}
\examples{
\dontrun{
session <- ssh_connect("servername")
submit_split("c:/yft/hessian")
}

}
