% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian-package.R
\docType{package}
\name{hessian-package}
\alias{hessian-package}
\alias{hessian}
\title{Parallel Hessian Computations and Diagnostics}
\description{
Organize parallel computations of the Hessian matrix in MULTIFAN-CL and run
diagnostics.
}
\details{
\emph{Main interface:}
\tabular{ll}{
  \code{\link{function}} \tab do stuff
}
}
\references{
\url{https://github.com/PacificCommunity/ofp-sam-hessian}
}
\seealso{
\pkg{hessian} is used in combination with the \pkg{condor} and \pkg{FLR4MFCL}
packages.
}
\author{
Arni Magnusson.
}
