% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian_merge.R
\name{hessian_merge}
\alias{hessian_merge}
\title{Hessian Merge}
\usage{
hessian_merge(working.dir, top.dir = "condor_hessian", ...)
}
\arguments{
\item{working.dir}{local directory containing Hessian subdirectories.}

\item{top.dir}{top directory on Condor submitter machine containing parallel
Hessian computations.}

\item{\dots}{passed to \code{\link{condor_download}}.}
}
\value{
Data frame showing downloaded jobs.
}
\description{
Gather \verb{.hes} files from parallel computations into one directory.
}
\examples{
\dontrun{
session <- ssh_connect("servername")
hessian_merge("c:/x/yft/hessian")
}

}
