% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian_merge.R
\name{hessian_merge}
\alias{hessian_merge}
\title{Hessian Merge}
\usage{
hessian_merge(model, quiet = FALSE)
}
\arguments{
\item{quiet}{whether to suppress messages.}

\item{working.dir}{local directory containing Hessian subdirectories.}
}
\value{
Directory containing the \verb{.hes} files, ready to merge.
}
\description{
Gather \verb{.hes} files from parallel computations into one directory.
}
\examples{
\dontrun{
hessian_merge("z:/yft/2023/model_runs/hessian/diagnostic")
}

}
