% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian_standalone.R
\name{hessian_standalone}
\alias{hessian_standalone}
\title{Hessian Stand-Alone}
\usage{
hessian_standalone(original.dir, working.dir = basename(original.dir),
  overwrite = FALSE)
}
\arguments{
\item{original.dir}{directory containing a converged model run.}

\item{working.dir}{directory where Hessian run will be prepared.}

\item{overwrite}{whether to remove existing working directory.}
}
\value{
Filenames in working directory after preparation.
}
\description{
Prepare working directory for Hessian stand-alone calculation, starting from
a converged model run.
}
\note{
The term \emph{stand-alone} refers to a calculation of the Hessian that
starts from a converged model.

A stand-alone calculation takes a shorter time to run than a
\emph{monolithic} run, where an extended \verb{doitall.sh} script first fits
a model and then calculates the Hessian.

A stand-alone calculation takes a longer time to run than a \emph{parallel}
Hessian calculation, where computations are run simultaneously on many cores.
}
\examples{
\dontrun{
hessian_standalone("z:/yft/2023/model_runs/diagnostic")
}

}
