% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian_submit.R
\name{hessian_submit}
\alias{hessian_submit}
\title{Hessian Submit}
\usage{
hessian_submit(working.dir, top.dir = "condor_hessian", ...)
}
\arguments{
\item{working.dir}{local directory containing Hessian subdirectories.}

\item{top.dir}{top directory on Condor submitter machine for running parallel
Hessian computations.}

\item{\dots}{passed to \code{\link{condor_submit}}.}
}
\value{
Type of output.
}
\description{
Submit parallel Hessian computations to Condor.
}
\examples{
\dontrun{
hessian_submit("c:/x/yft/hessian")
}

}
